Separate PyFD from PyFDAlgorithm
